requirements.txt
matplotlib==3.8.0
numpy==1.23.0
pyyaml==6.0.2
tqdm==4.67.1
torch>=2.1.0,<2.3.0
torchvision>=0.16.0,<0.18.0
tensorboard==2.0.0
protobuf==3.20.3